MODLOGIN-116: reduce HttpClient creation to one time
Moved the login API's `HttpClient` creation to `InitAPIs`. The client is stored in the vertx context shared by the APIs. A static `HttpClient` singleton would not work well for testing without modifying the security model to reset the static. Cleaning up the client seemed challenging without refactoring much of the code.

A couple things worth noting:
1. I moved the `lookup.timeout` `HttpClientOptions` as well. However, I see we are using the same value for both the connect timeout and the idle timeout. The value itself defaults to `1000` which would suggest milliseconds and the connect timeout is in ms, however, the idle timeout unit defaults to seconds. This may not be the intended usage here and should probably be looked into at some point.
1. Also, at some point, `ConfigurationServiceImpl` and `PasswordStorageServiceImpl` should be profiled to see if they are leaking `HttpClient` objects (it looks like they would). But, they use the `EventBus`, so it is not entirely clear.