Has most of Julian's latest but not all
I'm going to leave the renaming of the cookie keys for now since I beleve Adam's okapi
code uses the current names. I'd like to deploy to the rancher env first with the 4 branches and
the existing names first.

This code also has SameSite and Domain which I'm still thinking about and want to try testing
in chrome's dev tools after we deploy to rancher.