Address Adam's concern, bump minor version
Adam's concern about the cookie parser failing if clients don't present an AT key value
in the cookie let me to remove passing the AT through to mod-authtoken entirely. There's no
reason to pass it through since mod-at does nothing with it. Its important to note
that the authn/refresh endpoint can receive cookies that either have or don't have AT
keys and values. When requests come from browser clients, they will have both, which is why
this didn't get noticed earlier. But if the client is a backend module, it will
likely be that the cookie doesn't have an AT, which is perfectly fine, since it might
be expired anyway. The key thing to notice is that RT validation doesn't require
anything but the RT.

I also bumped the minor version based on other comments. It's not a breaking change
(yet). When we turn off the old endpoint, it might be.