Adds X-Okapi-Token back to request between mod-login and mod-authtoken
With the previous change I misunderstood what kind of token this is. It's a ModuleToken not an AccessToken
and therefore can and should be used to authorize communication between the two modules. There's no risk
of it expring. I added a comment to remind myself and perhaps point this out to others.